pyproject.toml
[project]
name = "news-notification-system"
version = "0.1.0"
description = "ニュース収集・要約・LINE通知システム"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "line-bot-sdk>=3.5.0",
    "google-genai>=1.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "mypy>=1.5.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.1.1",
    "pytest>=8.3.5",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle Errors
    "W",    # pycodestyle Warnings
    "F",    # Pyflakes
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"